+++ pom.xml
@@ -19,6 +19,14 @@
 	</properties>
 
 	<dependencies>
+
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>4.12</version>
+			<scope>test</scope>
+		</dependency>
+
 		<dependency>
 			<groupId>com.github</groupId>
 			<artifactId>jsqltools</artifactId>
@@ -76,7 +84,7 @@
 
 	<build>
 		<finalName>LookDbs</finalName>
-	<resources>
+		<resources>
 			<resource>
 				<directory>src/main/resources</directory>
 				<includes>
@@ -112,7 +120,7 @@
 			</resource>
 		</resources>
 		<plugins>
-		<plugin>
+			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-dependency-plugin</artifactId>
 				<version>2.8</version>
@@ -134,7 +142,7 @@
 				</executions>
 			</plugin>
 
-	 		<plugin>
+			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-jar-plugin</artifactId>
 				<version>3.0.2</version>
@@ -148,14 +156,14 @@
 					</archive>
 				</configuration>
 				<executions>
-                    <execution>
-                        <phase>compile</phase>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-			</plugin> 
+					<execution>
+						<phase>compile</phase>
+						<goals>
+							<goal>jar</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
 		</plugins>
 	</build>
